fix(daemon): split checkPeerUID into platform-specific files
The original PR-176 placed unix.GetsockoptUcred + unix.SO_PEERCRED in
pkg/daemon/ipc.go without a build tag. These syscalls are Linux-only;
macOS CI failed at 'go vet' with:
  pkg/daemon/ipc.go:479:24: undefined: unix.GetsockoptUcred
  pkg/daemon/ipc.go:479:71: undefined: unix.SO_PEERCRED

Split checkPeerUID into:
- ipc_peercred_linux.go  — SO_PEERCRED + Ucred (original logic)
- ipc_peercred_darwin.go — LOCAL_PEERCRED + GetsockoptXucred (BSD equivalent)
- ipc_peercred_other.go  — no-op fallback for unsupported platforms

Tested locally on macOS: go build + go test pass.
